requirements.txt
RPi.GPIO==0.7.1
Pillow>=9.0.0
spidev>=3.5
